+++ source/Examples/Resources/DnsServerSetting/2-DnsServerSetting_RemoteNode_Config.ps1
@@ -53,7 +53,6 @@ Configuration DnsServerSetting_RemoteNode_Config
             IsSlave            = $true
             RoundRobin         = $true
             LocalNetPriority   = $true
-            SecureResponses    = $true
             NoRecursion        = $false
             BindSecondaries    = $false
             StrictFileParsing  = $false